Split up test_area.py
It was too long - heading towards 1000 lines.
Split out test_checksum_api.py and test_validation_api.py
